Tell telemetry location manager to allow bg updates if required
In iOS 9, for background location updates while suspended it is no
longer sufficient to only set the UIBackgroundModes key with the
location value in the host app info.plist. You must also set
the value of `allowsBackgroundLocationUpdates` in CLLocationManager
to YES. This allows application logic to opportunistically opt in to
background updates when a user enables relevant app features that
make use of such updates.

This commit teaches the MGLMapboxEvents class to set this flag with
a guard for past (and future) cases where it may not make sense. Note
that this change alone won't cause background telemetry reporting to
take fully function -- it just makes it possible on iOS 9 when the
other related configuration is done correctly.